getPosts

+++ src/actions/user.action.ts
@@ -2,6 +2,7 @@
 
 import prisma from "@/lib/prisma";
 import { auth, currentUser } from "@clerk/nextjs/server";
+import { revalidatePath } from "next/cache";
 
 export async function syncUser() {
   try {
@@ -44,42 +45,44 @@ export async function getUserByClerkId(clerkId: string) {
         select: {
           followers: true,
           following: true,
-          posts: true
-        }
-      }
-    }
-  })
+          posts: true,
+        },
+      },
+    },
+  });
 }
 
 export async function getDbUserId() {
-  const {userId: clerkId} = await auth();
-  if (!clerkId) throw new Error("Unauthorized")
-  
-    const user = await getUserByClerkId(clerkId)
-    
-    if (!user) throw new Error("User not found");
+  const { userId: clerkId } = await auth();
+  if (!clerkId) return null;
 
-    return user.id
+  const user = await getUserByClerkId(clerkId);
+
+  if (!user) throw new Error("User not found");
+
+  return user.id;
 }
 export async function getRandomUsers() {
   try {
     const userId = await getDbUserId();
 
+    if (!userId) return []
+
     // get 3 random users exclude ourselves & users that we already follow
     const randomUsers = await prisma.user.findMany({
       where: {
         AND: [
-          {NOT: {id: userId}},
+          { NOT: { id: userId } },
           {
             NOT: {
               followers: {
                 some: {
-                  followerId: userId
-                }
-              }
-            }
-          }
-        ]
+                  followerId: userId,
+                },
+              },
+            },
+          },
+        ],
       },
       select: {
         id: true,
@@ -89,14 +92,14 @@ export async function getRandomUsers() {
         _count: {
           select: {
             followers: true,
-          }
-        }
+          },
+        },
       },
       take: 3,
-    })
-    return randomUsers
+    });
+    return randomUsers;
   } catch (error) {
-    console.log("Error fetching random users" , error)
+    console.log("Error fetching random users", error);
     return [];
   }
 }
@@ -105,49 +108,51 @@ export async function toggleFollow(targetUserId: string) {
   try {
     const userId = await getDbUserId();
 
-    if (userId === targetUserId) throw new Error("You cannot follow yourself")
+    if (!userId) return;
 
-      const existingFollow = await prisma.follows.findUnique({
+    if (userId === targetUserId) throw new Error("You cannot follow yourself");
+
+    const existingFollow = await prisma.follows.findUnique({
+      where: {
+        followerId_followingId: {
+          followerId: userId,
+          followingId: targetUserId,
+        },
+      },
+    });
+
+    if (existingFollow) {
+      //unfollow
+      await prisma.follows.delete({
         where: {
           followerId_followingId: {
             followerId: userId,
-            followingId: targetUserId
-          }
-        }
-      })
-
-      if(existingFollow) {
-        //unfollow
-        await prisma.follows.delete({
-          where: {
-            followerId_followingId: {
-              followerId: userId,
-              followingId: targetUserId
-            }
-          }
-        })
-      } else {
-        // follow
-        await prisma.$transaction([
-          prisma.follows.create({
-            data: {
-              followerId: userId,
-              followingId: targetUserId
-            }
-          })
-        ]),
+            followingId: targetUserId,
+          },
+        },
+      });
+    } else {
+      // follow
+      await prisma.$transaction([
+        prisma.follows.create({
+          data: {
+            followerId: userId,
+            followingId: targetUserId,
+          },
+        }),
+      ]),
         prisma.notification.create({
           data: {
             type: "FOLLOW",
-            userId:targetUserId, // user being followed
-            creatorId: userId // user following
-          }
-        })
-      }
-
-      return {success: true}
+            userId: targetUserId, // user being followed
+            creatorId: userId, // user following
+          },
+        });
+    }
+    revalidatePath("/https://codervai.vercel.app/");
+    return { success: true };
   } catch (error) {
-    console.log("Error in toggelFollow ", error)
-    return {success: false, error: "Error in toggling follow"}
+    console.log("Error in toggelFollow ", error);
+    return { success: false, error: "Error in toggling follow" };
   }
 }
